backend/requirements.txt
# Core Framework
fastapi==0.110.0
uvicorn[standard]==0.29.0

# Database
sqlalchemy==2.0.29
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication
python-jose[cryptography]==3.3.0
bcrypt>=4.0.0
python-multipart==0.0.9

# AI / Embeddings
sentence-transformers==2.7.0
faiss-cpu==1.8.0
numpy==1.26.4

# Resume Parsing
pdfplumber==0.11.0
spacy==3.7.4

# Utility
pydantic==2.7.0
pydantic-settings==2.2.1
email-validator>=2.0.0
python-dotenv==1.0.1
httpx==0.27.0

# Supabase (optional - for real-time features)
supabase>=2.0.0
